# docs/swagger.yaml
basePath: /
definitions:
  controller._ResponseLogin:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/controller.ResCode'
        description: 业务响应状态码
      data:
        description: 响应数据
        type: string
      message:
        description: 提示信息
        type: string
      ok:
        type: boolean
    type: object
...
  controller._ResponseToAssignRoleList:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/controller.ResCode'
        description: 业务响应状态码
      data:
        $ref: '#/definitions/model.ResponseToAssignRole'
      message:
        description: 提示信息
        type: string
      ok:
        type: boolean
    type: object
  controller._ResponseUserInfo:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/controller.ResCode'
        description: 业务响应状态码
      data:
        $ref: '#/definitions/model.ResponseUserInfo'
      message:
        description: 提示信息
        type: string
      ok:
        type: boolean
    type: object
  controller._ResponseUserList:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/controller.ResCode'
        description: 业务响应状态码
      data:
        $ref: '#/definitions/model.ResponseUserList'
      message:
        description: 提示信息
        type: string
      ok:
        type: boolean
    type: object
  controller._ResponseUserSingUP:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/controller.ResCode'
        description: 业务响应状态码
      data:
        type: string
      message:
        description: 提示信息
        type: string
      ok:
        type: boolean
    type: object
  controller.ResCode:
    enum:
    - 200
    - 201
    - 202
    - 203
    - 204
    - 205
    - 206
    - 207
    - 208
    type: integer
    x-enum-varnames:
    - CodeSuccess
    - CodeInvalidParam
    - CodeUserExist
    - CodeUserNotExist
    - CodeInvalidPassword
    - CodeServerBusy
    - CodeInvalidToken
    - CodeNeedLogin
    - CodeMenuNodeExist
  controller.ResponseData:
    properties:
      code:
        $ref: '#/definitions/controller.ResCode'
      data: {}
      message: {}
      ok:
        type: boolean
    type: object
...
  model.ParamDoAssignRole:
    properties:
      roleIdList:
        items:
          type: integer
        type: array
      userId:
        type: integer
    type: object
...
  model.ParamUserLogin:
    properties:
      password:
        description: 密码，不能为空
        type: string
      username:
        description: 用户名，不能为空
        type: string
    required:
    - password
    - username
    type: object
  model.ParamUserSignUp:
    properties:
      name:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - name
    - password
    - username
    type: object
  model.ParamUserUpdate:
    properties:
      name:
        description: 用户昵称
        type: string
      username:
        description: 用户姓名
        type: string
    required:
    - name
    - username
    type: object
...
  model.ResponseToAssignRole:
    properties:
      allRolesList:
        items:
          $ref: '#/definitions/model.Role'
        type: array
      assignRoles:
        items:
          $ref: '#/definitions/model.Role'
        type: array
    type: object
  model.ResponseUser:
    properties:
      ID:
        type: integer
      createTime:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      roleName:
        type: string
      updateTime:
        type: string
      username:
        type: string
    type: object
  model.ResponseUserInfo:
    properties:
      avatar:
        type: string
      buttons:
        items:
          type: string
        type: array
      name:
        type: string
      roles:
        items:
          type: string
        type: array
      routes:
        items:
          type: string
        type: array
    type: object
  model.ResponseUserList:
    properties:
      current:
        type: integer
      pages:
        type: integer
      records:
        items:
          $ref: '#/definitions/model.ResponseUser'
        type: array
      size:
        type: integer
      total:
        type: integer
    type: object
  model.Role:
    properties:
      ID:
        type: integer
      createTime:
        type: string
      id:
        type: integer
      remark:
        type: string
      roleName:
        type: string
      updateTime:
        type: string
    type: object
...
host: 127.0.0.1:10086
info:
  contact:
    email: 929000201@qq.com
    name: WangJian
    url: https://github.com/wangjian3306
  description: 硅谷甄选项目后端
  title: vue3_admin 项目接口文档
  version: "1.0"
paths:
  /admin/acl/index/info:
    get:
      consumes:
      - application/json
      description: 获取用户信息
      parameters:
      - description: 用户 Token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponseUserInfo'
      security:
      - ApiKeyAuth: []
      summary: 用户信息接口
      tags:
      - 后台登录与菜单管理
  /admin/acl/index/login:
    post:
      consumes:
      - application/json
      description: 处理用户登录请求
      parameters:
      - description: 用户登录参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/model.ParamUserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponseLogin'
      security:
      - ApiKeyAuth: []
      summary: 用户登录接口
      tags:
      - 后台登录与菜单管理
  /admin/acl/index/logout:
    post:
      consumes:
      - application/json
      description: 处理用户登出
      parameters:
      - description: 用户 Token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: 用户登出接口
      tags:
      - 后台登录与菜单管理
...
  /admin/acl/user/{page}/{limit}:
    get:
      consumes:
      - application/json
      description: 获取用户列表
      parameters:
      - description: 用户 Token
        in: header
        name: token
        required: true
        type: string
      - description: 当前页码
        in: path
        name: page
        required: true
        type: integer
      - description: 每页记录数
        in: path
        name: limit
        required: true
        type: integer
      - description: 用户名
        in: query
        name: username
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponseUserList'
      security:
      - ApiKeyAuth: []
      summary: 获取用户分页列表
      tags:
      - 用户管理
  /admin/acl/user/doAssignRole:
    post:
      consumes:
      - application/json
      description: 为用户分配角色
      parameters:
      - description: 用户 Token
        in: header
        name: token
        required: true
        type: string
      - description: 用户角色
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/model.ParamDoAssignRole'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: 为用户分配角色
      tags:
      - 用户管理
  /admin/acl/user/remove/{id}:
    delete:
      consumes:
      - application/json
      description: 处理删除用户请求
      parameters:
      - description: 用户 Token
        in: header
        name: token
        required: true
        type: string
      - description: 用户 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: 删除用户接口
      tags:
      - 用户管理
  /admin/acl/user/save:
    post:
      consumes:
      - application/json
      description: 处理用户新增请求
      parameters:
      - description: 用户登录参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/model.ParamUserSignUp'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponseUserSingUP'
      security:
      - ApiKeyAuth: []
      summary: 用户新增接口
      tags:
      - 用户管理
  /admin/acl/user/toAssign/{adminId}:
    get:
      consumes:
      - application/json
      description: 获取用户角色分配息
      parameters:
      - description: 用户 Token
        in: header
        name: token
        required: true
        type: string
      - description: 用户 ID
        in: path
        name: adminId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponseToAssignRoleList'
      security:
      - ApiKeyAuth: []
      summary: 用户角色分配接口
      tags:
      - 用户管理
  /admin/acl/user/update:
    put:
      consumes:
      - application/json
      description: 处理更新用户请求
      parameters:
      - description: 用户 Token
        in: header
        name: token
        required: true
        type: string
      - description: 用户登录参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/model.ParamUserUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: 更新用户接口
      tags:
      - 用户管理
...
swagger: "2.0"